[AIP-21] Move contrib.operators.azure_cosmos_operator airflow.contrib.operators.azure_cosmos_operator
