+++ scripts/ci/pylint_todo.txt
@@ -32,7 +32,7 @@
 ./airflow/providers/microsoft/azure/hooks/wasb.py
 ./airflow/providers/microsoft/azure/operators/adls_list.py
 ./airflow/providers/microsoft/azure/operators/azure_container_instances.py
-./airflow/contrib/operators/azure_cosmos_operator.py
+./airflow/providers/microsoft/azure/operators/azure_cosmos.py
 ./airflow/contrib/operators/dingding_operator.py
 ./airflow/contrib/operators/discord_webhook_operator.py
 ./airflow/providers/apache/druid/operators/druid.py
